Exp_5/requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu118
# torch==2.2.2+cu118
# torchvision==0.17.2+cu118
# torchaudio==2.2.2+cu118
# xformers==0.0.25.post1+cu118
# pip install torch==1.12.0+cu113 torchvision==0.13.0+cu113 torchaudio==0.12.0 --extra-index-url https://download.pytorch.org/whl/cu113
# pip install torch==1.12.1+cu113 torchvision==0.13.1+cu113 torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu113
# xformers==0.0.16    # https://blog.csdn.net/wangning6252820/article/details/134056100
# wget https://anaconda.org/xformers/xformers/0.0.16/download/linux-64/xformers-0.0.16-py310_cu11.3_pyt1.12.1.tar.bz2
# conda install xformers-0.0.16-py310_cu11.3_pyt1.12.1.tar.bz2
# python -m xformers.info
omegaconf==2.3.0
accelerate==0.28.0 
einops==0.7.0
opencv_python==4.9.0.80
scipy==1.12.0
ftfy==6.2.0
regex==2023.12.25
python-dateutil==2.9.0.post0
timm==0.9.16
pytorch-lightning==1.9.5 # only for loading pretrained sd weight   # pip install pytorch-lightning==2.2.1 torch限制，原来是
tensorboard==2.16.2 # for tensorboard event visualization
protobuf==4.25.3 # for tensorboard
lpips==0.1.4
facexlib==0.3.0
gradio==4.43.0
polars==1.12.0
torchsde==0.2.6
bitsandbytes==0.44.1

# requirements for llava
transformers==4.37.2
tokenizers==0.15.1
sentencepiece==0.1.99

# requirements for ram
fairscale==0.4.4


#  某个推理时候的问题
# cd /root/anaconda3/envs/diffbir/lib/python3.10/site-packages/bitsandbytes/
# cp libbitsandbytes_cuda117.so libbitsandbytes_cpu.so


# Could not find the bitsandbytes CUDA binary at 
# PosixPath('/root/anaconda3/envs/diffbir/lib/python3.10/site-packages/bitsandbytes/libbitsandbytes_cuda113.so
Pillow==9.5.0
